+++ packages/signals/lib/signals.dart
@@ -7,18 +7,22 @@ export 'src/signals.dart'
         effect,
         batch,
         untracked,
+        EffectCleanup,
+        Computed,
+        Signal,
+        ValueSignal,
         MutableSignal,
         ReadonlySignal;
+export 'src/future_signal.dart' show futureSignal, FutureSignal;
+export 'src/stream_signal.dart' show streamSignal, StreamSignal;
+export 'src/list_signal.dart' show listSignal, ListSignal;
+export 'src/map_signal.dart' show mapSignal, MapSignal;
+export 'src/set_signal.dart' show setSignal, SetSignal;
+export 'src/iterable_signal.dart' show iterableSignal, IterableSignal;
+export 'src/connect.dart' show connect, Connect;
 export 'src/extensions/stream.dart';
 export 'src/extensions/future.dart';
 export 'src/extensions/list.dart';
 export 'src/extensions/map.dart';
 export 'src/extensions/set.dart';
 export 'src/extensions/iterable.dart';
-export 'src/future_signal.dart' show futureSignal;
-export 'src/stream_signal.dart' show streamSignal;
-export 'src/list_signal.dart' show listSignal;
-export 'src/map_signal.dart' show mapSignal;
-export 'src/set_signal.dart' show setSignal;
-export 'src/iterable_signal.dart' show iterableSignal;
-export 'src/connect.dart' show connect;