+++ packages/signals/lib/src/iterable_signal.dart
@@ -1,7 +1,7 @@
 import 'signals.dart';
 
 /// A [Signal] that holds a [Iterable].
-class IterableSignal<E> extends Signal<Iterable<E>> implements Iterable<E> {
+class IterableSignal<E> extends ValueSignal<Iterable<E>> implements Iterable<E> {
   /// Creates a [IterableSignal] with the given [value].
   IterableSignal(super.value, {super.debugLabel});